Changed TileMap.GetTile() to take a Vector2<int>. Conversion shouldnt be TileMaps job.

+++ src/CLGEngine/TileMap.h
@@ -19,6 +19,6 @@ public:
     CLGEngine::Vector2<int> offset = {0, 0}; // We currently need this to grab global Tile location. Will have to set from GameManager for now.
 
     TileMap(MapData data);
-    wchar_t GetTile(CLGEngine::Vector2<float> coords);
+    wchar_t GetTile(CLGEngine::Vector2<int> coords);
     void SetMap(MapData map);
 };
